Minor clean-up, scss vars, move roster functions to utilities.ts
