+++ src/assets/scss/_variables.scss
@@ -25,6 +25,8 @@ $secondaryGray: #858585;
 $lightPlaceholderGray: #757575;
 $darkPlaceholderGray: #B6B6B6;
 
+$lightLabelGray: #9E9E9E;
+
 // Export for javascript
 :export {
   yuxuanBlue: $yuxuanBlue;
