+++ src/components/Requirements.vue
@@ -48,6 +48,7 @@ import SubRequirement from '@/components/SubRequirement.vue';
 import { BaseRequirement as Requirement, CourseTaken, SingleMenuRequirement } from '@/requirements/types';
 import { RequirementMap, computeRequirements, computeRequirementMap } from '@/requirements/reqs-functions';
 import { AppUser, AppMajor, AppMinor, AppSemester, FirestoreSemesterCourse } from '@/user-data';
+import { getRostersFromLastTwoYears } from '@/utilities';
 
 const functions = firebase.functions();
 
@@ -85,7 +86,7 @@ export default Vue.extend({
   },
   mounted() {
     this.recomputeRequirements();
-    this.rostersFromLastTwoYears = this.getRostersFromLastTwoYears();
+    this.rostersFromLastTwoYears = getRostersFromLastTwoYears();
   },
   data() : Data {
     return {
@@ -259,53 +260,6 @@ export default Vue.extend({
       return `<b>This is your Requirements Bar <img src="${clipboard}"class = "newSemester-emoji-text"></b><br>
           <div class = "introjs-bodytext">To ease your journey, we’ve collected a list of course
           requirements based on your college and major :)</div>`;
-    },
-    getCurrentType() {
-      let currentType;
-      const currentMonth = new Date().getMonth();
-      if (currentMonth === 0) {
-        currentType = 'WI';
-      } else if (currentMonth <= 4) {
-        currentType = 'SP';
-      } else if (currentMonth <= 7) {
-        currentType = 'SU';
-      } else {
-        currentType = 'FA';
-      }
-      return currentType;
-    },
-    getCurrentYearSuffix() {
-      // If current year is 2020, get string '20'
-      const currentYear = new Date().getFullYear();
-      return currentYear.toString().substring(2);
-    },
-    getRostersFromLastTwoYears() {
-      // If current roster is FA20, get all rosters thru FA18
-      const currentType = this.getCurrentType();
-      const currentYearSuffix = this.getCurrentYearSuffix();
-
-      const currentRoster = currentType + currentYearSuffix;
-      const numRosters = 9; // Number of most recent rosters we want
-      const mostRecentRosters = [currentRoster];
-      let roster = currentRoster;
-      while (mostRecentRosters.length < numRosters) {
-        // Go backwards until we hit 2 years and add to list of rosters
-        let rosterType = roster.substring(0, 2);
-        let rosterYear = roster.substring(2);
-        if (rosterType === 'FA') {
-          rosterType = 'SU';
-        } else if (rosterType === 'SU') {
-          rosterType = 'SP';
-        } else if (rosterType === 'SP') {
-          rosterType = 'WI';
-        } else if (rosterType === 'WI') {
-          rosterType = 'FA';
-          rosterYear = (parseInt(rosterYear, 10) - 1).toString();
-        }
-        roster = rosterType + rosterYear;
-        mostRecentRosters.push(roster);
-      }
-      return mostRecentRosters;
     }
   }
 });
